--- realman/RMC-LM/requirements.txt
@@ -1,1 +1,0 @@
-mcp==1.9.4
